+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Changelog
 - Pass options to .read (include, exclude, noRaw, onlyRaw)
 - Read unsynchronisation & dataLengthIndicator of frame header (v2.4.0)
+- Skip extended header if present
 
 ### Changed
 - Frames are now build/read by a frame builder definition instead of the manual programmed way